requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary
dotenv==0.9.9
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.4.4
pika==1.3.2
platformdirs==4.3.6
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
twilio==9.6.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.1.3
yarl==1.20.0
